Move methods from Helper.cs to BoreholeGeometryExtensions.cs
